Add files via upload
Added BoardModel. It's useable, but I will also add a way of counting how many X or Os are on the board. Not sure where to place the 'check if the game has been won' method, but once at least 5 turns have passed (3 for one player, 2 for the other) we should start running the check. (NOTE: or we could just run it every time, regardless of how many turns have passed)